requirements.txt
Authlib==1.0.1
cachelib==0.9.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
Flask==2.2.2
Flask-Reuploaded==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2==1.9.0.post1
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
Pygments==2.9.0
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
tqdm==4.56.0
urllib3==1.26.11
uWSGI==2.0.20
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
